backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "sql-agent"
version = "0.1.0"
description = "SQL Agent with LangGraph"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.9",
    "pydantic>=2.7.4",
    "pydantic-settings>=2.1.0",
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "psycopg2-binary>=2.9.9",
    "asyncpg>=0.29.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "structlog>=24.1.0",
    "langsmith>=0.5.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
